Move non-camera specific test from /test/imagetests to /test
