+++ test/MisplacedExifTest.php
@@ -22,7 +22,7 @@
  * Boston, MA 02110-1301 USA
  */
 
-namespace Pel\Test\imagetests;
+namespace Pel\Test;
 
 use lsolesen\pel\Pel;
 use lsolesen\pel\PelJpeg;
@@ -39,7 +39,7 @@ class MisplacedExifTest extends TestCase
         Pel::clearExceptions();
         Pel::setStrictParsing(false);
         // Image contains non-EXIF APP1 section ahead of the EXIF one
-        $jpeg = new PelJpeg(dirname(__FILE__) . '/misplaced-exif.jpg');
+        $jpeg = new PelJpeg(dirname(__FILE__) . '/broken_images/misplaced-exif.jpg');
         // Assert we just have loaded correct file for the test
         $this->assertNotInstanceOf('\lsolesen\pel\PelExif', $jpeg->getSection(PelJpegMarker::APP1));
