+++ test/Tags1Test.php
@@ -34,7 +34,7 @@ class Tags1Test extends TestCase
     {
         Pel::clearExceptions();
         Pel::setStrictParsing(true);
-        $jpeg = new PelJpeg(dirname(__FILE__) . '/test-tags-1.jpg');
+        $jpeg = new PelJpeg(dirname(__FILE__) . '/images/test-tags-1.jpg');
 
         $exif = $jpeg->getExif();
         $this->assertInstanceOf('\lsolesen\pel\PelExif', $exif);
