+++ test/broken_images/misplaced-exif.jpg
